Ticker
I added a news ticker to the top of the news page. I couldn't get the
border-radius to match the rest of the pages, and I also don't
understand the CSS of the forum enough to get that in the same div, but
if one of you guys could do that (and also figure out what to put in the
ticker) that'd be awesome.